+++ crates/subspace-farmer/Cargo.toml
@@ -35,6 +35,7 @@ serde_json = "1.0.64"
 subspace-archiving = { version = "0.1.0", path = "../subspace-archiving" }
 subspace-solving = { version = "0.1.0", path = "../subspace-solving" }
 subspace-core-primitives = { version = "0.1.0", path = "../subspace-core-primitives" }
+subspace-rpc-primitives = { version = "0.1.0", path = "../subspace-rpc-primitives" }
 thiserror = "1.0.24"
 tokio = { version = "1.11.0", features = ["macros", "rt-multi-thread"] }
